重构单词查询为公共 Hook
